api/requirements.txt
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bottle==0.12.25
Brotli==1.1.0
browser-cookie3==0.19.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
curl_cffi==0.6.2
defusedxml==0.7.1
duckduckgo_search==5.3.0
fastapi==0.110.2
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
g4f==0.3.0.3
greenlet==3.0.3
gTTS==2.5.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
loguru==0.7.2
lz4==4.3.3
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
nest-asyncio==1.6.0
orjson==3.10.1
pillow==10.3.0
platformdirs==4.2.0
plyer==2.1.0
proxy_tools==0.1.0
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
PyExecJS==1.5.1
python-socks==2.4.4
pythonnet==3.0.3
pywebview==5.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
starlette==0.37.2
tinycss2==1.2.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
webencodings==0.5.1
Werkzeug==3.0.2
win32-setctime==1.1.0
yarl==1.9.4
flasgger==0.9.7.1

bcrypt==4.1.2

PyJWT==2.9.0
